% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as-css-latex}
\alias{as-css-latex}
\alias{as_css}
\alias{as_latex}
\title{Export units as CSS/LaTeX}
\usage{
as_css(x)

as_latex(x)
}
\arguments{
\item{x}{An object of class \verb{units`` or }mixed_units`}
}
\value{
A character vector.
}
\description{
These functions throw an error if \code{units(x)} is not a valid CSS/LaTeX unit.
}
\examples{

unloadNamespace("cssunits")
unloadNamespace("units")
library(cssunits)
install_css_units()

x <- set_css_units(1:3, "pt")
as_css(x)

}
