% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{css-latex-unit-databases}
\alias{css-latex-unit-databases}
\alias{css_unit_db}
\alias{latex_unit_db}
\alias{latex_macro_db}
\title{Data frames of CSS and LaTeX units}
\format{
An object of class \code{data.frame} with 27 rows and 6 columns.

An object of class \code{data.frame} with 15 rows and 6 columns.

An object of class \code{data.frame} with 14 rows and 6 columns.
}
\usage{
css_unit_db

latex_unit_db

latex_macro_db
}
\description{
The package exports three data frames:  \code{css_unit_db},
\code{latex_unit_db} and \code{latex_macro_db}.
}
\details{
The format is not the same as the udunits2 database. Rows are:
\itemize{
\item \code{symbol}: symbol, e.g. "pt"
\item \code{from_unit}: unit on which the new unit is based, or \code{NA}
\item \code{new}: is unit new or already defined in udunits2?
\item \code{remove}: list-column of conflicting units to remove from udunits2
\item \code{type}: one of percent, length, angle, duration, frequency, resolution
}

Some units are redefined by "cssunits" so as to use a new
symbol. For example, "arc_degree" is present in udunits2, but it is replaced
by a new "deg" unit here.
}
\references{
\itemize{
\item \url{https://www.w3.org/TR/css3-values/}
\item \url{https://raw.githubusercontent.com/debian-tex/texlive-bin/master/texk/web2c/pdftexdir/pdftex.web}
}
}
\author{
David Hugh-Jones \href{mailto:davidhughjones@gmail.com}{davidhughjones@gmail.com}
}
\keyword{data}
\keyword{datasets}
