% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssunits-package.R
\docType{data}
\name{css_unit_db}
\alias{css_unit_db}
\alias{latex_unit_db,}
\alias{latex_macro_db}
\alias{latex_unit_db}
\title{Data frame of CSS and LaTeX units}
\description{
The format is not the same as the udunits2 database. Rows are:
}
\details{
\itemize{
\item \code{symbol}: symbol, e.g. "pt"
\item \code{from_unit}: unit on which the new unit is based, or \code{NA}
\item \code{new}: is unit new or already defined in udunits2?
\item \code{remove}: list-column of conflicting units to remove from udunits2
\item \code{type}: one of percent, length, angle, duration, frequency, resolution
}

Some units are redefined by "cssunits" so as to use a new
symbol. For example, "arc_degree" is present in udunits2, but it is replaced
by a new "deg" unit here.
}
\references{
\itemize{
\item \url{https://www.w3.org/TR/css3-values/}
\item \url{https://raw.githubusercontent.com/debian-tex/texlive-bin/master/texk/web2c/pdftexdir/pdftex.web}
}
}
\author{
David Hugh-Jones \href{mailto:davidhughjones@gmail.com}{davidhughjones@gmail.com}
}
\keyword{data}
