% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssunits-package.R
\docType{package}
\name{cssunits-package}
\alias{cssunits}
\alias{cssunits-package}
\title{cssunits: Units for CSS and LaTeX}
\description{
Modifies the units package to provide standard CSS and/or LaTeX
    units, including lengths, angles, times, frequencies etc.
}
\details{
The "cssunits" package uses the "units" package to define a set of valid
CSS and/or LaTeX units. Some defaults from the udunits2 database are removed
so that they can be replaced by more appropriate units. For example, "pt"
refers to points and not to pints.

To use it, run\preformatted{library(units)
library(cssunits)
install_css_units()
# or
install_latex_units()
}
}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
\keyword{internal}
