% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ex_half_em}
\alias{ex_half_em}
\title{Define ex = 0.5em}
\usage{
ex_half_em(on = TRUE)
}
\arguments{
\item{on}{\code{TRUE}or \code{FALSE} to switch behaviour on or off.}
}
\description{
Define ex = 0.5em
}
\details{
By default the package makes no assumption about the value of the length unit
ex, which is font-dependent. Use this function to make an ex equal to
0.5em.
}
\examples{
library(units)
install_css_units()
ex_half_em()
set_units(1, "em") + set_units(1, "ex")

}
