% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from-css-latex}
\alias{from-css-latex}
\alias{from_css}
\alias{from_latex}
\title{Import units from CSS or LaTeX character vectors.}
\usage{
from_css(x)

from_latex(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
An object of class "mixed_units".
}
\description{
Import units from CSS or LaTeX character vectors.
}
\examples{

unloadNamespace("cssunits")
unloadNamespace("units")
library(cssunits)
install_css_units()

from_css(c("3em", "14px"))

}
