% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_css_units}
\alias{install_css_units}
\title{Install units for CSS}
\usage{
install_css_units()
}
\description{
Install units for CSS
}
\details{
Some units will be removed from the standard database, including
"pt" (pint), "t" (tonne), "pi" (the mathematical constant), "h" (hour),
"rem" (unit of radiation).
}
\section{Removing}{

These functions install incompatible versions of some units, e.g. CSS points
are not the same as LaTeX points. They are also not idempotent: calling
them twice will probably generate an error.
To return to the pristine udunits2 database, use\preformatted{unloadNamespace("cssunits")
unloadNamespace("units")
library(units)
}
}

\examples{

unloadNamespace("cssunits")
unloadNamespace("units")
library(cssunits)
library(units)
install_css_units()
set_units(1, "in") + set_units(72, "pt")

}
