% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_latex_units}
\alias{install_latex_units}
\title{Install length units for LaTeX}
\usage{
install_latex_units(macros = FALSE)
}
\arguments{
\item{macros}{Boolean. Install some common LaTeX length macros such
as \verb{\\\\textwidth}?}
}
\description{
Install length units for LaTeX
}
\details{
Some units will be removed from the standard udunits database, including
"pt" (pint), "t" (tonne), "d" (day), "cc" (cubic centimeters), and
"u" (unit of atomic mass).
}
\section{Removing}{

These functions install incompatible versions of some units, e.g. CSS points
are not the same as LaTeX points. They are also not idempotent: calling
them twice will probably generate an error.
To return to the pristine udunits2 database, use\preformatted{unloadNamespace("cssunits")
unloadNamespace("units")
library(units)
}
}

\examples{

library(units)
install_latex_units()
set_units(1, "in") + set_units(72.27, "pt")

}
