% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{set-units}
\alias{set-units}
\alias{set_css_units}
\alias{set_latex_units}
\title{Set CSS/LaTeX units}
\usage{
set_css_units(x, value, ..., mode = units_options("set_units_mode"))

set_latex_units(x, value, ..., mode = units_options("set_units_mode"))
}
\arguments{
\item{x, value, ..., mode}{See \code{\link[units:set_units]{units::set_units()}}}
}
\value{
An object of class units. Throws an error if the unit specification
is not recognized.
}
\description{
These functions check that units are legitimate CSS/LaTeX units, then
call \code{\link[units:set_units]{units::set_units()}}.
}
\examples{

unloadNamespace("cssunits")
unloadNamespace("units")
library(cssunits)

install_css_units()
set_css_units(1, "px", mode = "standard")
set_css_units(1, px)
# catching errors:
try(set_css_units(1, pox))

}
