% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{units-from-strings}
\alias{units-from-strings}
\alias{from_css}
\alias{from_latex}
\title{Create units from CSS or LaTeX dimensions}
\usage{
from_css(x)

from_latex(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
An object of class "mixed_units".
}
\description{
Create units from CSS or LaTeX dimensions
}
\examples{

unloadNamespace("cssunits")
unloadNamespace("units")
library(cssunits)
install_css_units()

from_css(c("3em", "14px"))

}
