% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_css_units}
\alias{install_css_units}
\title{Install units for CSS}
\usage{
install_css_units()
}
\description{
Install units for CSS
}
\details{
Source: https://www.w3.org/TR/css3-values/, 10 May 2020.

Some units will be removed from the standard database, including
"pt" (pint), "t" (tonne), "pi" (the mathematical constant), "h" (hour),
"rem" (unit of radiation).
}
\examples{

library(units)
install_css_units()
set_units(72, "pt") + set_units(1, "in")
}
