% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_latex_units}
\alias{install_latex_units}
\title{Install length units for LaTeX}
\usage{
install_latex_units(macros = FALSE)
}
\arguments{
\item{macros}{Boolean. Install some common LaTeX length macros such
as \verb{\\\\textwidth}?}
}
\description{
Install length units for LaTeX
}
\details{
Values are sourced from
https://raw.githubusercontent.com/debian-tex/texlive-bin/master/texk/web2c/pdftexdir/pdftex.web.

Some units will be removed from the standard udunits database, including
"pt" (pint), "t" (tonne), "d" (day), "cc" (cubic centimeters), and
"u" (unit of atomic mass).
}
\examples{

library(units)
install_latex_units()
set_units(72, "pt") + set_units(1, "in")
}
